+++ src/main/java/woowacourse/chatting/repository/MemberRepository.java
@@ -1,4 +1,4 @@
-package woowacourse.chatting.Repositorty;
+package woowacourse.chatting.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
