+++ src/main/java/woowacourse/chatting/service/MemberDetailsService.java
@@ -5,7 +5,7 @@ import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
-import woowacourse.chatting.Repositorty.MemberRepository;
+import woowacourse.chatting.repository.MemberRepository;
 
 @Service
 @RequiredArgsConstructor
