+++ src/test/java/woowacourse/chatting/repository/MemberRepositoryTest.java
@@ -1,4 +1,4 @@
-package woowacourse.chatting.Repositorty;
+package woowacourse.chatting.repository;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -9,7 +9,6 @@ import woowacourse.chatting.domain.Member;
 import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 @DataJpaTest
 class MemberRepositoryTest {
